requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
requests==2.32.3
pyyaml==6.0.2
pydantic==2.9.2

